refactor(gateway): extract ApiProxyConfig + ProxyBackendConfig (GW-2 S5)
Also replaces the shared default_true() helper with a local
default_circuit_breaker_enabled() inside api_proxy.rs for the only
field outside config.rs that used it. Strictly identical behavior.
